API

API documentation

Mortar2D.calculate_normals — Function.

calculate_normals(elements::Dict{Int, Vector{Int}},
                  element_types::Dict{Int, Symbol},
                  X::Dict{Int, Vector{Float64})

Given elements, element types and node locations, calculate nodal normals by first calculating normal directions for each element and then averaging them in nodes. As a result we get unique normal direction defined to each node.

Example

X = Dict(1 => [7.0, 7.0], 2 => [4.0, 3.0], 3 => [0.0, 0.0])
elements = Dict(1 => [1, 2], 2 => [2, 3])
element_types = Dict(1 => :Seg2, 2 => :Seg2)
normals = calculate_normals(elements, element_types, X)

# output

Dict{Int64,Array{Float64,1}} with 3 entries:
  2 => [0.707107, -0.707107]
  3 => [0.6, -0.8]
  1 => [0.8, -0.6]

source

Mortar2D.project_from_master_to_slave — Function.

Find the projection of a master node xm, to the slave surface with nodes (xs1, xs2), in direction of slave surface normal defined by (ns1, ns2).

source

Mortar2D.project_from_slave_to_master — Function.

Find the projection of a slave node xs, having normal vector ns, onto master elements with nodes (xm1, xm2).

Mathematics

References

Yang2005

source

Mortar2D.calculate_segments — Function.

Given slave surface elements, master surface elements, nodal coordinates and normal direction on nodes of slave surface elements, calculate contact segments.

Returns a dictionary, where key is slave element id and values is a list of master elements giving contribution to that slave elements and xi-coordinates of slave side element.

source

Mortar2D.calculate_mortar_matrices — Function.

Given segmentation, calculate mortar matrices De and Me.

source

Mortar2D.calculate_mortar_assembly — Function.

Given data, calculate projection matrix P.

source

Index

Mortar2D.calculate_mortar_assembly

Mortar2D.calculate_mortar_matrices

Mortar2D.calculate_normals

Mortar2D.calculate_segments

Mortar2D.project_from_master_to_slave

Mortar2D.project_from_slave_to_master